requirements.txt
# ================================================
# Bonat AI Agent - Python Dependencies
# ================================================
# Install with: pip install -r requirements.txt

# === Web Framework & API ===
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

# === AI & Language Models ===
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langchain-community>=0.3.0
langgraph>=0.2.0
openai==1.102.0

# === Data Processing ===
pandas==2.2.3
numpy>=1.21.0,<2.0.0

# === Database Connectors ===
sqlalchemy>=2.0.0
pymongo==4.14.1
psycopg2-binary==2.9.9
PyMySQL>=1.1.0
mysql-connector-python>=8.0.0

# === Authentication & Security ===
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1

# === Configuration & Validation ===
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
PyYAML>=6.0  # For YAML semantic model loading

# === Template Engine ===
Jinja2>=3.1.0

# === Vector Database & RAG ===
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0
rank-bm25==0.2.2  # For hybrid search (BM25 keyword matching)

# === Additional Dependencies ===
# Add any other packages your deployment needs below
sqlparse==0.5.3
openpyxl>=3.1.0  # For Excel report generation in stress testing
psutil>=5.9.0  # For memory/CPU diagnostics